tests(multi-stamp): moves scan validation out from shared examples as it can be overriden by child components
